Allow parsing buffers, add has_anims to return.
If you're using io instead of love.fs to read the files or want to
reduce read calls, this will be useful. has_anims is intended to be used
to see if you should call iqm.load_anims (working on this right now)
after loading the mesh.

load_anims will be a separate call on the grounds that you do not always
want all animations in a file, and animation data can (and should) be
stored completely independently of your mesh data.